requirements.txt
pynput==1.7.6
Pillow==10.2.0
PyAutoGUI==0.9.54
pyserial==3.5